LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/container-queries/anchored-fallback-implicit-any-expected.txt


FAIL @container anchored(fallback: <position-area>) querying single axis with implicit 'any' assert_equals: expected "yes" but got ""
